If output path set to AOSP don't sync overlays if none

+++ scripts/generate-vendor.sh
@@ -1273,12 +1273,14 @@ if [[ "$AOSP_ROOT" != "" ]]; then
   echo "[*] Vendor blobs copied to '$AOSP_ROOT/vendor/$VENDOR_DIR'"
 
   # Vendor overlays are always under separate directories so it's safe to force update
-  mkdir -p "$AOSP_ROOT/vendor_overlay/$VENDOR_DIR"
-  rsync -arz --delete "$OUTPUT_DIR/vendor_overlay/$VENDOR_DIR/" "$AOSP_ROOT/vendor_overlay/$VENDOR_DIR/" || {
-    echo "[-] rsync failed"
-    abort 1
-  }
-  echo "[*] Vendor overlays copied to '$AOSP_ROOT/vendor_overlay/$VENDOR_DIR'"
+  if [[ "$OVERLAYS_DIR" != "" ]]; then
+    mkdir -p "$AOSP_ROOT/vendor_overlay/$VENDOR_DIR"
+    rsync -arz --delete "$OUTPUT_DIR/vendor_overlay/$VENDOR_DIR/" "$AOSP_ROOT/vendor_overlay/$VENDOR_DIR/" || {
+      echo "[-] rsync failed"
+      abort 1
+    }
+    echo "[*] Vendor overlays copied to '$AOSP_ROOT/vendor_overlay/$VENDOR_DIR'"
+  fi
 fi
 
 abort 0
